feat: ViewController구현 - ListViewController 클래스생성 및  TableView DataSource구현

+++ GyroData/ListModels/Measurement.swift
@@ -11,5 +11,5 @@ struct Measurement: Hashable {
     
     let sensor: Sensor
     let date: Date
-    let value: Int
+    let value: Double
 }
